+++ R/srppphist-data.R
@@ -172,8 +172,15 @@
 #' end of the use period ('expirationDeadline') are given in the respective
 #' columns of the `products` table.
 #'
+#' As of version 1.0.0.9004, parallel imports are included in the table, and
+#' the Swiss registration number ("Eigenössische Zulassungsnummer") is given
+#' in the column `chNbr`. For regular products, this column lists the W-Number
+#' in the format W-XXXX. For parallel import, the format is D-XXXX for a product
+#' from the German market, F-XXXX for a product from the French market and so on.
+#'
 #' @docType data
-#' @format tibble A table of all unique combinations of P-Number, W-Number
+#' @format tibble A table of all unique combinations of P-Number, W-Number,
+#' Swiss registration number
 #' and product name over all years, and the columns 'earliest' and 'latest',
 #' indicating the earliest and latest year of occurrence of that combination.
 #' Finally, the columns 'categories_de', 'categories_fr', and 'categories_it'
@@ -187,7 +194,7 @@
 #' year. In particular, a number of products with such comments are present
 #' in the register used as representative for 2011.
 #' @examples
-#' print(srppp_products[1:6], n = Inf)
+#' print(srppp_products[1:7], n = Inf)
 "srppp_products"
 
 #' Table of product ingredients extracted from all years starting 2011